more

+++ public/python_scripts/gemini_flash_text_fromhtml13.py
@@ -54,22 +54,19 @@ fn_s_l = sorted(glob.glob('/media/dan/ssd2/tmp/enhanced_htmls/2023*.html'))[:2]
 
 html_fn_s = fn_s_l[0]
 txt_fn_s = html_fn_s.replace('.html','.txt')
-
-with open(html_fn_s, 'r') as myhf:
-    html_s = myhf.read()
-
-invoke_d = {'html_syntax': html_s}
-
-llm_s = llm_s_l[0]
-
-myllm_model = ChatGoogleGenerativeAI(model=llm_s)
-
-chain = prompt | myllm_model | parser
-
-text_fromhtml_s = chain.invoke(invoke_d)
-
-with open(txt_fn_s, 'w') as mytf:
-    mytf.write(text_fromhtml_s)
+if os.path.exists(txt_fn_s):
+    'Done already'
+else:
+    'I will extract case info into: ' + txt_fn_s
+    with open(html_fn_s, 'r') as myhf:
+        html_s = myhf.read()
+    invoke_d = {'html_syntax': html_s}
+    llm_s = llm_s_l[0]
+    myllm_model = ChatGoogleGenerativeAI(model=llm_s)
+    chain = prompt | myllm_model | parser
+    text_fromhtml_s = chain.invoke(invoke_d)
+    with open(txt_fn_s, 'w') as mytf:
+        mytf.write(text_fromhtml_s)
 
 'done'
